moved io tests to their own file
